// duman-cpms/src/model/Account.ts
export class Account {
    activity_certificate_pdf: string;
    authorized_signatures_list_pdf: string;
    birthplace: string;
    board_management: string;
    city: string;
    created_date: string;
    data_id: number;
    district: string;
    email_addr: string;
    fax: string;
    firm_address: string;
    firm_full_name: string;
    firm_key_contact_personnel: string;
    firm_type: string;
    identity_copies: string;
    partnership_structure_identity_copies: string;
    phone_number: string;
    sector: string;
    tax_department: string;
    tax_return_pdf: string;
    taxpayer_number: string;
    web_url: string;
}